Add empty messages and headers
