+++ src/routes/MyRequests/components/MyRequestsPage/MyRequestsPage.js
@@ -1,24 +1,18 @@
 import React from 'react';
-import { format } from 'date-fns';
 import { Helmet } from 'react-helmet';
-import { useFirestore, useFirestoreDoc, useUser } from 'reactfire';
-import { Link } from 'react-router-dom';
+import { useFirestore, useUser, useFirestoreCollection } from 'reactfire';
 import Paper from '@material-ui/core/Paper';
 import Container from '@material-ui/core/Container';
 // import Divider from '@material-ui/core/Divider';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
-import Grid from '@material-ui/core/Grid';
-import Card from '@material-ui/core/Card';
 // import Paper from '@material-ui/core/Paper';
-import Alert from '@material-ui/lab/Alert';
-import Chip from '@material-ui/core/Chip';
-import FinancialAssistanceIcon from '@material-ui/icons/AttachMoney';
-import GroceryIcon from '@material-ui/icons/ShoppingBasket';
 import { makeStyles } from '@material-ui/core/styles';
 import { REQUESTS_COLLECTION } from 'constants/collections';
-import { allCategoryMap } from 'constants/categories';
+import { Link } from 'react-router-dom';
+import { NEW_REQUEST_PATH } from 'constants/paths';
 import styles from './MyRequestsPage.styles';
+import RequestCard from '../RequestCard';
 
 const useStyles = makeStyles(styles);
 
@@ -26,28 +20,21 @@ function MyRequestsPage() {
   const classes = useStyles();
   const firestore = useFirestore();
   const user = useUser();
-
-  const requestRef = firestore
+  const userRequestsRef = firestore
     .collection(REQUESTS_COLLECTION)
     .where('createdBy', '==', user.uid);
-  const requestsSnap = useFirestoreDoc(requestRef);
+  const openRequestsRef = userRequestsRef.where('status', '<', 20);
+  const closedRequestsRef = userRequestsRef.where('status', '==', 20);
+  const openRequestsSnap = useFirestoreCollection(openRequestsRef);
+  const closedRequestsSnap = useFirestoreCollection(closedRequestsRef);
 
-  function handleCopyNeedLink(id) {
-    const el = document.createElement('textarea');
-    document.body.appendChild(el);
-    el.value = `${window.location.origin}/requests/${id}`;
-    el.select();
-    document.execCommand('copy');
-    document.body.removeChild(el);
-  }
-
-  if (!requestsSnap.docs) {
+  if (!openRequestsSnap.docs && !closedRequestsSnap.doc) {
     return (
       <Container>
         <Helmet>
           <title>Requests Not Found</title>
         </Helmet>
-        <Paper className={classes.paper} data-test="request-not-found">
+        <Paper className={classes.paper} data-test="requests-not-found">
           Request Not Found
         </Paper>
       </Container>
@@ -59,109 +46,37 @@ function MyRequestsPage() {
       <Helmet>
         <title>My Requests</title>
       </Helmet>
-      {requestsSnap &&
-        requestsSnap.docs.map((requestSnap) => (
-          <Card
-            key={requestSnap.id}
-            className={classes.cards}
-            data-test="request-card"
-            data-test-id={requestSnap.id}>
-            <Container maxWidth="lg" className={classes.TaskContainer}>
-              <Grid container>
-                <Grid item xs={9}>
-                  <Typography variant="caption" gutterBottom>
-                    ADDED{' '}
-                    {requestSnap.get('createdAt') &&
-                      typeof requestSnap.get('createdAt').toDate ===
-                        'function' &&
-                      format(requestSnap.get('createdAt').toDate(), 'p - PPPP')}
-                  </Typography>
-                  <Typography
-                    variant="h5"
-                    className={classes.Needs}
-                    gutterBottom>
-                    {requestSnap.get('needs') &&
-                      Object.keys(requestSnap.get('needs')).map((item) => (
-                        <React.Fragment key={item}>
-                          {allCategoryMap[item] ? (
-                            <Chip
-                              variant="outlined"
-                              icon={
-                                item === 'grocery-pickup' ? (
-                                  <GroceryIcon />
-                                ) : null
-                              }
-                              label={allCategoryMap[item].shortDescription}
-                            />
-                          ) : (
-                            <Alert severity="error">
-                              Could not find &apos;{item}&apos; in all category
-                              map.
-                            </Alert>
-                          )}
-                        </React.Fragment>
-                      ))}
-                    {requestSnap.get('needFinancialAssistance') && (
-                      <Chip
-                        variant="outlined"
-                        icon={<FinancialAssistanceIcon />}
-                        label="Need financial assistance"
-                      />
-                    )}
-                  </Typography>
-                </Grid>
-
-                {parseInt(requestSnap.get('immediacy'), 10) > 5 && (
-                  <Grid item xs={3}>
-                    <img
-                      align="right"
-                      src="/taskIcon.png"
-                      width="50px"
-                      height="50px"
-                      alt="Urgent"
-                      title="Urgent"
-                    />
-                  </Grid>
-                )}
-                <Grid
-                  className={classes.DetailsButton}
-                  align="right"
-                  item
-                  xs={12}>
-                  {navigator.share ? (
-                    <Button
-                      color="primary"
-                      variant="outlined"
-                      onClick={() => {
-                        navigator.share({
-                          title: 'CV19 Assist Need Link',
-                          text: 'CV19 Assist Need Link',
-                          url: `${window.location.origin}/needs/${requestSnap.id}`,
-                        });
-                      }}>
-                      SHARE
-                    </Button>
-                  ) : (
-                    <Button
-                      color="primary"
-                      variant="outlined"
-                      onClick={() => handleCopyNeedLink(requestSnap.id)}>
-                      COPY LINK FOR SHARING
-                    </Button>
-                  )}{' '}
-                  <Button
-                    variant="contained"
-                    color="primary"
-                    disableElevation
-                    component={Link}
-                    to={`/requests/${requestSnap.id}`}>
-                    DETAILS...
-                  </Button>
-                </Grid>
-              </Grid>
-            </Container>
-          </Card>
-        ))}
+      <Typography variant="h5">Open Requests</Typography>
+      {openRequestsSnap && openRequestsSnap.docs.length ? (
+        openRequestsSnap.docs.map((requestSnap) => (
+          <RequestCard key={requestSnap.id} requestSnap={requestSnap} />
+        ))
+      ) : (
+        <Paper className={classes.paper}>
+          <Typography variant="body1" className={classes.emptyMessage}>
+            You do not currently have any open requests
+          </Typography>
+          <Button
+            variant="contained"
+            color="primary"
+            component={Link}
+            to={NEW_REQUEST_PATH}>
+            Create A New Request
+          </Button>
+        </Paper>
+      )}
+      <Typography variant="h5">Closed Requests</Typography>
+      {closedRequestsSnap && closedRequestsSnap.docs.length ? (
+        closedRequestsSnap.docs.map((requestSnap) => (
+          <RequestCard key={requestSnap.id} requestSnap={requestSnap} />
+        ))
+      ) : (
+        <Paper className={classes.paper}>
+          <Typography variant="body1">
+            Requests will appear here once they have been completed
+          </Typography>
+        </Paper>
+      )}
     </Container>
   );
 }
