+++ src/routes/MyRequests/components/MyRequestsPage/MyRequestsPage.styles.js
@@ -1,24 +1,12 @@
 export default (theme) => ({
-  cards: {
+  paper: {
+    textAlign: 'center',
     marginTop: theme.spacing(2),
-    marginBottom: theme.spacing(2),
-  },
-  TaskTitle: {
-    color: '#F4B7B4',
-  },
-  Needs: {
-    color: '#F4B7B4',
-    display: 'flex',
-    '& > *': {
-      marginTop: theme.spacing(0.5),
-      marginRight: theme.spacing(0.5),
-    },
-  },
-  TaskContainer: {
+    marginBottom: theme.spacing(10),
     paddingTop: theme.spacing(2),
     paddingBottom: theme.spacing(2),
   },
-  DetailsButton: {
-    paddingTop: theme.spacing(1),
+  emptyMessage: {
+    marginBottom: theme.spacing(4),
   },
 });
