+++ src/routes/MyRequests/components/RequestCard/index.js
@@ -1,0 +1,3 @@
+import RequestCard from './RequestCard';
+
+export default RequestCard;
